refactor: move UserSubmissionController and DTOs to challenge package
